Add a quicksort implementation
A standard, 2-way partition quicksort implementation. Choosing
a more sophisticated algorithm for computing the pivot is a worthy
experiment if the input is not already randomized (as it is
guaranteed to be right now). Additionally, a 3-way partition such as
the Bentley-McIlroy partition is another potential optimization and I
might consider adding it as a separate quicksort algorithm for comparison
in the future.